requirements.txt
customtkinter
keyring
pandas
undetected-chromedriver
selenium
Pillow
